HFP-3191 Add hack for fixing Panopto embed API bug
If there are multiple videos using the Panopto embed API the events from on the players are emitted to all of the players.
This causes the componentDidUpdate in InteractionContent.js to trigger every time one player tries to play a video, and the
pause logic would kick in.
This hack works around this by checking if both audio sources are a video, in this case we do not pause the video.
This hack should be removed when Panopto has fixed their embed API.